Update Calc_App.java
Last Updated June 7th, 2022:
Unlike before, the calculator will now allow for you to use the keyboard for as your input for the numbers and operators.

If anyone is willing, I've been having some trouble with the calculator's input with some operators. Whenever you type " * " as your operator (multiplication) it would input an 8 in the text field. It's because you have to click Shift and then click 8. The same thing happens with modulus but instead it inserts a 5 instead of 8. I'm not sure how to fix it, it would be much appreciated if anyone can help. Thanks.